backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.1
Werkzeug==3.0.3